[scm] Improve git error handling
Only fall back to a `rawTree` if git is not in $PATH or if the current
directory is not a git repository. Any other error likely represents a
bug in shac's handling of git and should cause the command to fail.

Change-Id: I069b6680c616ea00b6b21ed248840604533e646e
Reviewed-on: https://fuchsia-review.googlesource.com/c/shac-project/shac/+/826537
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Oliver Newman <olivernewman@google.com>
Reviewed-by: Marc-Antoine Ruel <maruel@google.com>