[engine] use threads
Add check contexts.

This adds more backtrace to the checks. We'll probably want to tidy the
errors afterwards since the information is now redundant.

This is the initial part to allow concurrent execution, which is the
foundation for shac's performant runtime characteristics.

Remove references to luci's errors implementation, as it is not
necessary anymore.

Change-Id: Ic463b56a634ac8c2ae9c58258402f77f3a63f82b
Reviewed-on: https://fuchsia-review.googlesource.com/c/shac-project/shac/+/829057
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Oliver Newman <olivernewman@google.com>
Fuchsia-Auto-Submit: Marc-Antoine Ruel <maruel@google.com>