[docs] Check that ctx doc in stdlib.star is up-to-date
Adds a check that compares the schema of the dummy `ctx` object in
stdlib.star to the actual `ctx` object that's passed into each check
callback to ensure the docs are in sync with the implementation.

Unfortunately I couldn't figure out a way to validate that the function
signatures are correct, but this is a helpful check to start with.

Change-Id: I9e8a89cb328a66ad9f631ffcc28f0ac3a01fc326
Reviewed-on: https://fuchsia-review.googlesource.com/c/shac-project/shac/+/830559
Fuchsia-Auto-Submit: Oliver Newman <olivernewman@google.com>
Commit-Queue: Oliver Newman <olivernewman@google.com>
Reviewed-by: Marc-Antoine Ruel <maruel@google.com>