[engine] Disallow lambda check callbacks without `name`
Previously at most one check in a project would be allowed to be
registered as a lambda function without an explicit name, and its name
would default to "lambda".

This is bad because it's confusing for lambda to only be allowed once,
and "lambda" is also probably not a useful name for a check, so we
should require that such checks explicitly set their names.

Change-Id: Ia9208ec000d0a66752412e0c4131623235351151
Reviewed-on: https://fuchsia-review.googlesource.com/c/shac-project/shac/+/834916
Fuchsia-Auto-Submit: Oliver Newman <olivernewman@google.com>
Reviewed-by: Marc-Antoine Ruel <maruel@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>