[engine] Check ctx.Err() if subprocess fails
Previously, if check A called `fail()` while check B was running a
subprocess, causing the subprocess to get canceled, the output would
make it look like it was actually check B that failed.

For example, adding a `fail("fail")` to `new_todos()` resulted in this:

  - new_todos (Success in 7ms): fail: fail
  - shadow (Success in 8ms): ctx.os.exec: command failed with exit code -1: ["go", "install", "golang.org/x/tools/go/analysis/passes/shadow/cmd/shadow@v0.7.0"]

Now the same code produces:

  - new_todos (Success in 9ms): fail: fail

and no status is reported for `shadow`.

It's confusing that a check still says "Success" if `fail()` is called
during it, I'll fix that separately.

Change-Id: I108a1062b339a9c2c781a149d7b3ada680e175e3
Reviewed-on: https://fuchsia-review.googlesource.com/c/shac-project/shac/+/836337
Fuchsia-Auto-Submit: Oliver Newman <olivernewman@google.com>
Reviewed-by: Marc-Antoine Ruel <maruel@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>