[engine] Add test for network sandboxing
Uses a bash script to attempt to query a local HTTP server. This is not
super portable and we'd prefer to use `go run`, but `go run` is too slow
to run from a shac check due to filesystem sandboxing blocking access to
`$GOCACHE`. Until we figure out a workaround, this provides a safety net
to prevent network sandboxing from regressing.

Change-Id: I7ef976025a05172d36f4e1933050906b10723a9c
Reviewed-on: https://fuchsia-review.googlesource.com/c/shac-project/shac/+/855376
Reviewed-by: Marc-Antoine Ruel <maruel@google.com>
Commit-Queue: Oliver Newman <olivernewman@google.com>