[engine] Don't ignore end_col if end_line is unset
If `end_col` is specified but `end_line` is not, we would previously
ignore `end_col` entirely and consider the finding to apply to the end
of the line.

Now we'll respect `end_col` in this case, and assume `end_line` should
be equal to `start_line`.

Change-Id: I57b42fd4ca9ae26f9bd72ba87b8d0de56b3efcb0
Reviewed-on: https://fuchsia-review.googlesource.com/c/shac-project/shac/+/878433
Fuchsia-Auto-Submit: Oliver Newman <olivernewman@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Marc-Antoine Ruel <maruel@google.com>