[engine] Add support for ignoring files in shac.textproto
It's useful to exclude files from ever being analyzed by shac. For
example, any formatting/linting issues in vendored third-party code
should generally be ignored (or fixed upstream) so that the code stays
in sync with upstream, rather than modifying the vendored version.

Ignored files are configured with a repeated `ignore` field in
shac.textproto. Each `ignore` element is a string that uses
fully-featured gitignore syntax.

We use the gitignore library from https://github.com/go-git/go-git to
parse the ignore fields, as it seemed to be the most well-maintained and
high-quality of the various available gitignore-parsing libraries for
Go. However, the library does contain a lot of things we don't need and
pulls in a number of dependencies, so at some point we should probably
reimplement the gitignore parsing ourselves and remove the dependency.

Change-Id: If0ad914a0754a1cb203eed620f3fd26c9a220fdb
Reviewed-on: https://fuchsia-review.googlesource.com/c/shac-project/shac/+/879285
Reviewed-by: Marc-Antoine Ruel <maruel@google.com>
Commit-Queue: Oliver Newman <olivernewman@google.com>