[engine] Add `formatter` arg to `shac.check()`
This allows checks to declare that they are formatters to be run by
`shac fmt`. Running only formatters is often significantly faster than
running all checks, and we expect formatters to be by far the most
common analyzers for which users will want to auto-apply fixes. So it
makes sense to carve out a special case for formatters.

The main downside is that it would be easy for a user to introduce a
formatting check and forget to mark it as `formatter=True`, and then be
surprised when it doesn't get run by `shac fmt`.

Alternatives considered:
- Automatically detect when a check is a formatter prior to running it
  without requiring any user-added annotations, e.g. if "fmt" or
  "format" is in the check name. But this is too magical and less
  discoverable than a `formatter` kwarg.
- Detect when a check may be a formatter *after* running it (e.g. if it
  emits a full-file finding with a replacement) and fail if the check is
  not marked as a formatter. This assumes that no non-formatter checks
  will ever emit full-file replacements (probably not true) and that
  formatter checks will always emit full-file replacements (probably not
  true, e.g. rustfmt has a diff mode).
- Provide a much more convenient way of writing formatting checks that
  distinguishes them from non-formatting checks (e.g. the check function
  has a different signature, or the `ctx` object has additional fields
  that are useful for writing formatting checks and can only be accessed
  by checks marked as formatters). This would still be nice to have and
  isn't mutually exclusive with a boolean `formatter` argument.

Change-Id: I0b8e8e96618d01c674ce67a1cd4ebcd77c995a3d
Reviewed-on: https://fuchsia-review.googlesource.com/c/shac-project/shac/+/881733
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Marc-Antoine Ruel <maruel@google.com>
Fuchsia-Auto-Submit: Oliver Newman <olivernewman@google.com>