[tests] Unzip contents of test module
To comply with licensing requirements, check in the unzipped contents of
the `ut` package instead of a zip file.

We can't use the normal Go vendoring process because it strips out
`_test.go` files and .git files, which are included in the go.sum hash
calculation and thus need to be available for the test.

Bug: 131750
Change-Id: I11333ff9bbb2a86fa53ef93da3f199a15c0307d2
Reviewed-on: https://fuchsia-review.googlesource.com/c/shac-project/shac/+/898278
Reviewed-by: Marc-Antoine Ruel <maruel@google.com>
Commit-Queue: Oliver Newman <olivernewman@google.com>
Fuchsia-Auto-Submit: Oliver Newman <olivernewman@google.com>