[engine] Add missing error return
This return isn't covered by any tests, and I don't see a way it could
be hit unless there's a significant bug in shac (e.g. non-function
Starlark objects are allowed to be registered as checks, which isn't the
case now) or in the starlark-go implementation. But it's best to be safe
in case such a bug does arise.

Change-Id: I3a87b0234ccdddeea583cd29a73df74b9323ab7a
Reviewed-on: https://fuchsia-review.googlesource.com/c/shac-project/shac/+/907842
Fuchsia-Auto-Submit: Oliver Newman <olivernewman@google.com>
Reviewed-by: Marc-Antoine Ruel <maruel@google.com>
Commit-Queue: Marc-Antoine Ruel <maruel@google.com>