[cli] Add --quiet flag to fix and fmt commands
Fuchsia's `fx format-code` command should not print any output on
success, so for it to be able to run `shac fmt`, there must be a way to
disable non-error output for `shac fmt`.

This should eventually apply to the `check` command as well for
consistency, but for now it's only needed on `fix` and `fmt`, and it was
easiest to do it this way because the output of `fix` and `fmt` comes
from a different codepath than that of `check`.

Change-Id: I6a1f895863523ed3e8ac3d0258f4ff2b07493c32
Reviewed-on: https://fuchsia-review.googlesource.com/c/shac-project/shac/+/911236
Fuchsia-Auto-Submit: Oliver Newman <olivernewman@google.com>
Reviewed-by: Anthony Fandrianto <atyfto@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>