[scm] Don't try to parse invalid ls-files output
If an error occurred on the `ls-files` command or a previous git
command, `g.err` will be set and will cause subsequent `g.run()` calls
to short-circuit and return an empty string. Trying to parse an empty
string using the ls-files parsing logic causes panics, so we must skip
it if an error occurred.

This was causing confusing panics when a check failed before `git
ls-files` was called for the first time.

Change-Id: Ice428fc83ecd9e7566e74c3ff85e03635676ff19
Reviewed-on: https://fuchsia-review.googlesource.com/c/shac-project/shac/+/916074
Reviewed-by: Anthony Fandrianto <atyfto@google.com>
Fuchsia-Auto-Submit: Oliver Newman <olivernewman@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>