[engine] Add check.with_args(...) method
This allows downstream repositories to override arguments to checks in
shared checks repositories, such as the path to an executable to run.

I also added a `with_name()` method, mainly to make it easier to test
`with_args()` (since it's not allowed to register multiple checks with
the same name) but it may also be useful for downstream users.

Change-Id: I43e50768bbf9c90d3c9aaa28ef447ef330c31d5c
Reviewed-on: https://fuchsia-review.googlesource.com/c/shac-project/shac/+/919559
Reviewed-by: Marc-Antoine Ruel <maruel@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Oliver Newman <olivernewman@google.com>