[checks] Set GOWORK=off
This ensures that Go commands always use the versions of libraries
specified in go.mod, and thus that are vendored in the /vendor
directory. Otherwise, when working in a Go workspace, Go commands would
try to use the version of each library from the workspace, which
wouldn't necessary be in the /vendor directory, and would fall back to
trying to download from the internet.

Also set `GOPROXY=off` for better error messages in the future - it
should now indicate better *why* it's trying to connect to the internet,
rather than trying to connect and failing.

Finally, update the `gosec` and `shadow` versions since the old versions
seem to be crashing on newer code.

Change-Id: I14b22ef9565a1181fc54d3c5cacf0573844358a5
Reviewed-on: https://fuchsia-review.googlesource.com/c/shac-project/shac/+/1058532
Fuchsia-Auto-Submit: Oliver Newman <olivernewman@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Danielle Kay <danikay@google.com>