[reporting] Split full-file diffs for readability
It's common for findings from code formatters to have suggested
replacements that cover the entire file. However, the resulting diff as
shown in Gerrit is not very readable because it shows the entire file
contents twice even when only a small number of lines would change.

Now we use a diffing library to split full-file diffs into small chunks,
so we don't show large unchanged parts of the file in Gerrit.

This is especially important now that we're migrating from Gerrit robot
comments to Gerrit findings for showing these results. Gerrit findings
present the suggested diff much more prominently than did robot
comments, so it's more important that the diff not be huge and
unhelpful.

Bug: 408229390
Change-Id: I39eb10348c8b9d466e1eebe8e1d688a26b2cae24
Reviewed-on: https://fuchsia-review.googlesource.com/c/shac-project/shac/+/1245364
Commit-Queue: Oliver Newman <olivernewman@google.com>
Reviewed-by: Ina Huh <ihuh@google.com>