[reporting] Use lower-level difflib output
Use the lower-level `GetGroupedOpCodes()` difflib function, which
returns a list of structs representing the diff output, instead of
using `GetUnifiedDiffString()` which returns a human-readable output
that then needs to be parsed into the Sarif format.

This greatly simplifies the conversion code and also fixes a bug caused
by the string parsing logic: if the old code didn't have a trailing
newline on the last line, and the new code added one, the diff would
squash the + and - lines into one because of the lack of a newline
(basically conflating newlines in source versus newlines to separate the
diff lines).

Bug: 408229390
Change-Id: Id6277274ccb089bb32cce620cc65c43f8997c599
Reviewed-on: https://fuchsia-review.googlesource.com/c/shac-project/shac/+/1248984
Fuchsia-Auto-Submit: Oliver Newman <olivernewman@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Ina Huh <ihuh@google.com>