[engine] Use result of slices.Compact
slices.Compact returns a value instead of mutating the input, so the old
code was not actually removing duplicates.

This only mattered when files were based as positional arguments to
`shac check`, which is not super common, and even less common that
someone would list the same file twice, hence why we didn't notice it
until it started being raised by the "lint" GitHub workflow.

I also updated the `govet` shac check to include the analyzer that was
catching this in the GitHub workflow.

Change-Id: I4a0eb6014d9d9dc1edc225ea60bf76d1fcc3f950